scripts/phase2/requirements_image_descriptor.txt
google-generativeai>=0.3.0
Pillow>=10.0.0